+++ scripts/performance/bundle-size/compute-bundle-size.js
