+++ scripts/performance/index.js
@@ -1,17 +1,12 @@
 const { runMain } = require('../lib/execution-utils')
-const { fetchPR, LOCAL_BRANCH } = require('../lib/git-utils')
-const { reportAsPrComment, LOCAL_COMMIT_SHA } = require('./report-as-a-pr-comment')
+const { reportAsPrComment } = require('./report-as-a-pr-comment')
 const { reportToDatadog } = require('./report-to-datadog')
-const { calculateBundleSizes } = require('./bundle-size/bundle-size-calculator')
-const { triggerSyntheticsTest, waitForSyntheticsTestToFinish } = require('./cpu-performance/synthetics-trigger')
-
-const RETRIES_NUMBER = 6 // Number of retries to get the synthetic test result
+const { calculateBundleSizes } = require('./bundle-size/compute-bundle-size')
+const { computeCpuPerformance } = require('./cpu-performance/compute-cpu-performance')
 
 runMain(async () => {
-  const prNumber = (await fetchPR(LOCAL_BRANCH)).number
-  const bundleSizes = calculateBundleSizes()
-  const resultId = await triggerSyntheticsTest(prNumber, LOCAL_COMMIT_SHA)
-  await waitForSyntheticsTestToFinish(resultId, RETRIES_NUMBER)
-  await reportToDatadog(bundleSizes)
-  await reportAsPrComment(bundleSizes)
+  const localBundleSizes = calculateBundleSizes()
+  await computeCpuPerformance()
+  await reportToDatadog(localBundleSizes)
+  await reportAsPrComment(localBundleSizes)
 })
